RFC: automatically create symlinks for channels
With this PR, Juliaup automatically creates symlinks in `~/.local/bin`
for each new channel and updates them accordingly. I have intentionally
disabled this on Windows for now since `symlink` is not supported there
and I am not sure if there even is a `PATH` equivalent on Windows.

This functionality follows other tools such as jill and also allows
users to avoid going through `julialauncher`, which can sometimes
complicate things, e.g. when running Julia through GDB.

One remaining question is whether we want to add an option or env var to
turn off this behavior.