support nightly channels
The way this works is that we mark which versions are nightly versions
in the versions db during the build process and then special-case those
when added as a channel or when updated. Right now, `juliaup update`
will always download new nightly tarballs, which could potentially be
wasteful if that's called again in quick succession, but I think for now
that should be a reasonable solution.

I have also gotten rid of a lot of code duplication in `build.rs`, which
should make adding new platforms easier in the future. I hope I didn't
miss anything there.

Closes #96