Search depot path for existing juliaup dir
Old behaviour would pick the first segment of JULIA_DEPOT_PATH to find the
juliaup dir. This commit finds the first segment with a juliaup/juliaup.json
file. If no such are found, it picks the first segment.
Empty segments are expanded to the default path (which is ~/.julia/juliaup).
If JULIA_DEPOT_PATH is not set, or it is set to the empty string, it will also
use the default path.