improvement: update Cargo.toml and Cargo.lock
Initially, I thought it was musl having issues
with openssl. The culprit was the workflow that
has not included the step to install openssl-dev
and pkgconfig which causes build failures.

The solution is to modify the workflow file and
add the necessary step as mentioned from a separate
pull/merge request.