Add support for nightlies (#601)
This change makes it possible to use nightly builds, using
the `nightly` or `nightly~ARCH` channels (where `ARCH`
indicates an architecture similar to those supported for
official releases). Adding such a channel will download the
latest nightly, query its version (e.g. `1.11.0-DEV.1351`),
and install that similarly to regular channels.

When updating channels using `juliaup update`, nightly
channels will periodically be updated by downloading the
latest nightly and comparing the version to the installed
one. The frequency of this operation is determined by the
`NightlyUpdateInterval` setting, which defaults to once
every day, and can be set with `juliaup config
nightlyupdateinterval` (passing a value in seconds).

---------

Co-authored-by: Tim Besard <tim.besard@gmail.com>