FreeBSD support
Note that this uses command line tar instead of tar-rs on FreeBSD.

Unlike most systems, we can be quite sure that `tar` exists when running
on FreeBSD, so it should be safe to call out to it and fail with a
descriptive error in the off chance that the host system is weird. This
allows us to avoid using the Rust tar library, which does not currently
support the POSIX-2008.1 format that is produced by default by BSD tar.