add flake.nix
This adds a flake.nix which can be used to build and
retrieve `juliaup` from the nix package manager.
It uses the `rustPlatform.buildRustPackage` function
from nixpkgs to build the Rust package, and the
23.11 pkg distribution.

Nix makes it easy to build and install `juliaup`,
either by running `nix build` or `nix run` in the project
directory or by referencing the flake from another nix configuration.