Include FreeBSD when generating versiondb
This splits out the versiondb part of PR 872. This ensures that any
updates to versiondb that happen while 872 is in development don't
require re-syncing, and it will mean that tests will pass locally.

For some reason, versions.json lists the FreeBSD triplet with the suffix
11.1, which does not reflect the version of FreeBSD used for building
and has not been supported for a long time. In fact, the current FreeBSD
binaries won't work on anything before 13.2. But in order to use
versions.json without some special-casing and/or refactoring, we'll just
retain the weird suffix.