Support overriding an override
Fixes #987.

Now we get a stderr message on `override add`, which indicates if an override is changed, new, or already exists.

If the channel is not installed, it gives the same error as julialauncher.

Please check carefully since I'm new to Rust. Especially not sure if the `iter_mut` pattern is idiomatic.